% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_resource.R
\name{add_resource}
\alias{add_resource}
\title{Add a hub resource}
\usage{
add_resource(package, fields)
}
\arguments{
\item{package}{A \code{character(1)} with the name of an existing hub package or
the path to a newly created (not yet submitted/accepted) hub package.}

\item{fields}{A named list with the data to be added to the
resource. Elements and content of the list are described in \code{?metadata}.}
}
\description{
This function adds a hub resource to the AH or EH package metadata.csv file.
It can be used while creating a new hub package or for adding data to an
existing package.
}
\examples{
## create a mock package with metadata.csv; these steps are usually
## done by `create_pkg()` ...
pkgdir <- tempfile()
metadata_path <- file.path(pkgdir, "inst", "extdata", "metadata.csv")
dir.create(dirname(metadata_path), recursive = TRUE)
writeLines(paste(names(metadata()), collapse = ","), metadata_path)


## create a metadata record
metadata <- metadata(
    Title = "ENCODE",
    Description = "a test entry",
    BiocVersion = "4.1",
    Genome = NA_character_,
    SourceType = "JSON",
    SourceUrl = "https://www.encodeproject.org",
    SourceVersion = "x.y.z",
    Species = NA_character_,
    TaxonomyId = NA_integer_,
    Coordinate_1_based = NA,
    DataProvider = "ENCODE Project",
    Maintainer = "tst person <tst@email.com>",
    RDataClass = "data.table",
    DispatchClass = "Rda",
    Location_Prefix = NA_character_,
    RDataPath = "ENCODExplorerData/encode_df_lite.rda",
    Tags = c("ENCODE", "Homo sapiens")
)

## add the record to the metadata
add_resource(pkgdir, metadata)

}
