% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Create and validate metadata}
\usage{
metadata(
  Title = character(1),
  Description = character(1),
  BiocVersion = package_version("0.0"),
  Genome = character(1),
  SourceType = character(1),
  SourceUrl = character(1),
  SourceVersion = character(1),
  Species = character(1),
  TaxonomyId = integer(1),
  Coordinate_1_based = NA,
  DataProvider = character(1),
  Maintainer = character(1),
  RDataClass = character(1),
  DispatchClass = character(1),
  Location_Prefix = character(1),
  RDataPath = character(1),
  Tags = character()
)
}
\arguments{
\item{Title}{character(1)}

\item{Description}{character(1).}

\item{BiocVersion}{The two-digit version of Bioconductor the
resource is being introduced into. Could be a character vector
\code{"4.1"} or an object created from \code{package_version()}, e.g.,
\code{package_version("4.1")}.}

\item{Genome}{character(1).}

\item{SourceType}{character(1).}

\item{SourceUrl}{character(1).}

\item{SourceVersion}{character(1). A description of the version of
the resource in the original source. Since source version may
not follow R / Bioconductor versioning practices, this field
is not restricted to a \code{package_version()} format.}

\item{Species}{character(1).}

\item{TaxonomyId}{integer(1).}

\item{Coordinate_1_based}{logical(1) are the genomic coordinates in
the resource 0-based, or 1-based? Use NA if genomic coordinates
are not present in the resource.}

\item{DataProvider}{character(1).}

\item{Maintainer}{character(1).}

\item{RDataClass}{character(1).}

\item{DispatchClass}{character(1).}

\item{Location_Prefix}{character(1).}

\item{RDataPath}{character(1).}

\item{Tags}{character() Zero or more tags describing the data.}
}
\description{
This functions makes a list of values that can be used to add as a
resource to a 'metadata.csv' file in a Hub package. The type of
each argument indicates the expected value, e.g., \code{Title = character(1)} indicates that it is expected to be a character
vector of length 1. See individual parameters for more information.
}
\examples{
metadata()

tst <- metadata(
    Title = "ENCODE",
    Description = "a test entry",
    BiocVersion = package_version("3.9"),
    Genome = NA_character_,
    SourceType = "JSON",
    SourceUrl = "https://www.encodeproject.org",
    SourceVersion = package_version("0.0"),
    Species = NA_character_,
    TaxonomyId = NA_integer_,
    Coordinate_1_based = NA,
    DataProvider = "ENCODE Project",
    Maintainer = "tst person <tst@email.com>",
    RDataClass = "data.table",
    DispatchClass = "Rda",
    Location_Prefix = NA_character_,
    RDataPath = "ENCODExplorerData/encode_df_lite.rda",
    Tags = c("ENCODE", "Homo sapiens")
)

}
