% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_resource.R
\name{publish_resource}
\alias{publish_resource}
\title{A function that publishes resource to the hub S3 bucket}
\usage{
publish_resource(path, object)
}
\arguments{
\item{path}{A \code{character(1)} path to the file or directory to be added to
the bucket.}

\item{object}{A \code{character(1)} to indicate how the file should be named on
the bucket.}
}
\description{
This function uses functionality from the aws.s3 package to put files or
directorieso on the Bioconductor's test hub S3 bucket. The user should
have already contacted the hubs maintainers at hubs@bioconductor.org to get
the necessary credentials to access the bucket. These credentials should be
delcared in the system environment prior to running this function.
}
\examples{
fl <- tempdir()

}
